requirements.txt
python-dotenv
fastapi
uvicorn
openai
requests
pillow
google-genai
huggingface-hub
sqlite3
transformers
torch
